Refactor logrotate config generation
This only changes /etc/logrotate-ops.conf if any relevant values in OVSDB
changed.

Currently supported values are in the "System" table, "logrotate_config"
column:
    period      one of hourly, daily, weekly, monthly, yearly
    maxsize     max size of log file, suffixed by k, M or G


Note: this removes the support for copying the rotated logs to a TFTP
server. It seemed like a half-baked implementation; if this is genuinely
needed I will add it back via logrotate's postrotate stanza, where this
belongs, and use curl to add support for HTTP, FTP etc. TFTP is the worst
choice for transferring stuff!

Signed-off-by: Bert Vermeulen <bert@biot.com>
Change-Id: I76bd4fac4df3b86f3fa6a5ad1979198fb201beb1