Don't hardcode additional layer names
.gitignore partially reads:

	yocto/halon

which causes a layer with that specific name to be ignored, but makes it
hard to add layers with different names, because it becomes necessary to
edit .gitignore for each new name, which obviously doesn't scale.

Instead, ignore any directory directly under yocto/ and unignore the two
directories that we know for a fact should be there.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@hpe.com>
Change-Id: I57b4ca737b1a75a4eab57511281bf79c5f930390