chg: dev: Add python-ast, Python Abstruct Syntax Tree package
Adding python-ast package in the generate-manifect-2.7.py, to
automatically generate the python-ast package in the
python-2.7-manifest.inc file.  With this change, we can include the
ast.py files in the final image by adding the python-ast package in the
appropriate RDEPENDS directive in the recipt file.  The first user of
this package will be ops-ansible.bb, as ansible runtime requires ast.py*
on the switch.  ops-ansible.bb review request will be sent separately.

Change-Id: Id70d2f0eb2bf4083fe99f48d9afc1107f346faa2
Signed-off-by: Kei Nohguchi <kei.nohguchi@hpe.com>