Have each machine configuration specify its switchd plugin
The whole game with:

	${@bb.utils.contains('MACHINE_FEATURES', 'ops-container', 'ops-switchd-container-plugin', '',d)}
	${@bb.utils.contains('MACHINE_FEATURES', 'broadcom', 'ops-switchd-opennsl-plugin', '',d)}

doesn't scale, as you end up with a whole sequence of those checks to
try to figure out one value.

The other way to do this properly is to use OVERRIDES, but even that
doesn't scale well, because you end up with a whole bunch of:

	SWITCHD_PLUGIN_PROVIDER_ops-container = ...
	SWITCHD_PLUGIN_PROVIDER_broadcom = ...
	SWITCHD_PLUGIN_PROVIDER_... = ...

in a single file.

Change-Id: I1c079720d52a9ae6d35137422aa8fde2502e3b3e
Signed-off-by: Marcelo E. Magallon <marcelo.magallon@hpe.com>