Move libthrift-*.so out of -dev package back into runtime
The standard do_package magic sees this libfoo-1.2.3.so file and goes
like "oh, development file! move!". Remove the pattern that moves
/usr/lib/*.so to the -dev package and add the specific file that needs
to be in the runtime package.

The whole libfoo-x.y.z.so thing started fifteen years ago when people
decided that designing stable APIs was too hard of a job. What happens
is that instead of this:

	gcc ... -lfoo

now you can say:

	gcc ... -lfoo-x.y.z

but since it's a pain to change this every time the minor version (y.z
or z depending your understanding of what a version means) changes, then
someone figured that it's clever to have libfoo.so pointing to
libfoo-x.y.z.so, which allows you to do this as well:

	gcc ... -lfoo

The whole thing should have been:

	libfoo-x.so.x.y.z
	libfoo-x.so -> libfoox.so.x.y.z

and that's it. In this way, I wouldn't have been chasing weird
workarounds fifteen years later.

Move the file back where it belongs.

Change-Id: I28be0eab8cc737127087c5449bd27c5b77016f3a
Signed-off-by: Marcelo E. Magallon <marcelo.magallon@hpe.com>