new: dev: Add ops-chef and libyaml recipe
Adding ops-chef.bb recipe file.  To the start, it has ruby as a runtime
dependency, so that we can have the ruby on the swtich.  This also
removes python-libyaml, which will be replaced by the libyaml, which
will be shared between python and ruby recipes.

Here is the snapshot of the /opt/chef as well as the ruby/gem versions
on the switch:

bash-4.3# ls -al /opt/chef
total 8
drwxr-xr-x 2 root root 4096 Dec  9 07:06 .
drwxr-xr-x 5 root root 4096 Dec  9 07:21 ..
bash-4.3# which ruby
/usr/bin/ruby
bash-4.3# ruby -v
ruby 1.9.3p429 (2013-05-15 revision 40747) [x86_64-linux]
bash-4.3# which gem
/usr/bin/gem
bash-4.3# gem -v
1.8.23
bash-4.3#

Change-Id: Ibd0df5d365288d748dc82f9b788a6a41aaf55099
Signed-off-by: Kei Nohguchi <kei.nohguchi@hpe.com>