chg: dev: ops.bb: Change Subsystem:interfaces to status category
This is a SRCREV change in ops.bb to enable the following feature:

This interfaces column of the Subsystem table should be status, instead
of the configuration, as this is not for the user to change.  This
causes the DC to remove all the reference, as DC doesn't provide the
subsystem informatino in the GET operation, which itself is an issue,
and when we PUT the result back to the system, DC will remove this
column, which ended up to clenaup the interfaces which is not referenced
by other tables, e.g. Bridge or VRF.

Change-Id: Ifa9baf218a0682011f85cf61df7883e987b2aa36
Depends-On: Ie2a5e295c102e81ebd8aded8514bca5ce6ab9038
Signed-off-by: Kei Nohguchi <kei.nohguchi@hpe.com>