new: dev: prometheus timeseries daemon
This adds the prometheus (http://prometheus.io) service to OpenSwitch.
It runes on :9090 and will later be exposed on http(s)://url/prometheus
when the web frontend design in OpenSwitch is stable.

The Prometheus node exporter is also included and runs on :9100.

Prometheus is configured to retain data for 24h or one device reboot,
whichever is sooner. Metrics are kept in tmpfs and are lost on reboot or
power events.

In addition to this golang support is added. Golang is needed to compile
Prometheus and the bare minimum to do so has been included. It is
untested how this will work with multiple architectures or actual cross
compilation.

Thanks to Marcelo Magallon for the golang base.

Change-Id: I009a3701dd07aa539b6f1b43342492a9af8e641b
Signed-off-by: Christian Svensson <blue@cmd.nu>