Add support for overriding platform YAML files
If the IMAGE_FEATURES contains "ops-p4", we need a different
configuration description file for the ports. Instead of just hardcoding
this into the base ops-hw-config.bb recipe, add support for a
PLATFORM_FLAVOR variant that provides overrides based on some decision
taken by the platform-specific .bbappend.

In this case, the genericx86-64 ops-hw-config.bbappend file looks at the
existance of the "ops-p4" feature in IMAGE_FEATURES, and if that's
present, then it sets the flavor to "P4". If there are p4-specific .yaml
files, the base recipe will prefer those. If not, it will use the base
ones.

Change-Id: Ia203ec2d2d07ad062b23cb6ff18cf6b1fb9d3c99
Signed-off-by: Marcelo E. Magallon <marcelo.magallon@hpe.com>