new: dev: PAM module to use CHAP for RADIUS auth
A patch is being added to the pam-radius recipe. This patch generates
the pam_radius_chap_auth.so module.
pam_radius_chap_auth.so has the same beahvior as pam_radius_auth.so
except that it encodes user password into the CHAP-Password
attribute when sending packet to the RADIUS server for user
authentication.

Change-Id: I01e03cb58cb84733bd38dec9a4aa469438227bc4
Signed-off-by: Rajeev Grover <rajeev_grover@hpe.com>