Change to use nginx webserver
This commit contains security tuning for making nginx suitable in
production. Changes are as follows.

Move ops-restd HTTPS to 8092 (disable when the code supports that).
ope-restd should in the future listen to only localhost.

Make /rest/, /login and /account map to localhost:8091 for ops-restd.

Remove dependency on ops-restd on ops-webui.

Ship 4096 Diffie-Hellman parameters that are not the same as the
standard ones used in nginx. This protects against Logjam.

Change-Id: I17258b6b152505254ba50141dca6fa0ae828a3f9
Signed-off-by: Christian Svensson <blue@cmd.nu>