chg: dev: Change to use nginx webserver
This commit contains security tuning for making nginx suitable in
production. Changes are as follows.

Change the user to run as www-data using capabilities to allow running
on <1024 ports. This requires kernel 4.3 and system 229.

Move ops-restd HTTPS to 8092 (disable when the code supports that).
ope-restd should in the future listen to only localhost.

Make /rest/, /login and /account map to localhost:8091 for ops-restd.

Remove dependency on lighttpd on ops-webui.

Ship 4096 Diffie-Hellman parameters that are not the same as the
standard ones used in nginx. This protects against Logjam.

Change-Id: I17258b6b152505254ba50141dca6fa0ae828a3f9
Signed-off-by: Christian Svensson <blue@cmd.nu>