add support for golang
This adds support for bootstrapping and building Go binaries. The
latest version is 1.6 and thus then one included here.

Since Go is written in Go you also need to bootstrap the 1.6 environment. For
this 1.4 is included as this version can be compiled with gcc.

If the developer is using binutils 2.26, which adds some new
relocations to amd64, we need to be using that to compile golang as well.
Therefore binutils 2.26 from poky is included.
See https://github.com/golang/go/issues/13114 for more info.

Thanks to Marcelo Magallon for the golang base.

Change-Id: I460c955bbed0db6e276ae079aa945cfd717f7e29
Signed-off-by: Christian Svensson <blue@cmd.nu>