Workaround for devenv_add problem
This:

	make devenv_add foo

causes a recipe "foo" with two or more Git sources in the SRC_URI to
have an alternates file pointing to a non-existent directory in the
"internal" repository (if you have more than one repository, you need to
build a nested directory structure, so at least one of the repos is
going to be "internal" to the parent repository).

The problem seems to be in the way that devtool calls bitbake tasks
(do_fetch specifically): even if it seems to be doing everything
correctly, it ends up with a repository where it's not possible to
switch branches or do much else because the alternates file still points
to a temporary copy which no longer exists after devtool finishes doing
its thing and this only happens for the "internal" repo.

Change-Id: I0133c769f66d83d21672745e6963ecdb9c3e4379
Signed-off-by: Marcelo E. Magallon <marcelo.magallon@hpe.com>