Workaround for devtool problem
This:

	make devenv_add foo

causes the internal repo to have an alternates file pointing to a
non-existent directory if the recipe "foo" is such that it specifies
more than one git repository in its SRC_URI. The problem seems to be in
the way that devtool calls bitbake tasks (do_fetch specifically), but
even if it seems to be doing everything correctly, it ends up with a
repository where it's not possible to switch branches or do much else.

Still looking for a better solution, but this is actually impacting a
few people now.

Change-Id: I0133c769f66d83d21672745e6963ecdb9c3e4379
Signed-off-by: Marcelo E. Magallon <marcelo.magallon@hpe.com>