new: dev: Add Ceedling (C Unit Test Framework) support
Ceedling is a Ruby based C Unit Test Framework. It uses Rake
as its build system.

This change adds the recipe for Ceedling and its dependency, Thor.
Also adds a Ceedling class, so projects can inherit from it. This
class will set up the environment for a project to work with the
framework. Note this required a refactor of the ruby class

Change-Id: I4f33ef8582a156cfbe52158730e6c6eb48a2f831
Signed-off-by: Pablo Barrantes <pbarrantes@hpe.com>