Make the testenv_run target bind to Docker volume
Tags: chg, dev, wip

With the current code coverage implementation, the way to retrieve
the coverage files from the Docker container was to bind the local
directory with the Docker container using Docker volumes.
This permits for the process in the container to read the gcno files
and then to write the gcda on process exit.

This change uses the binds argument that the Modular framework
reads from a json file and parses into -v (volume) arguments
to Docker when starting the container.
The change also configures the legacy test framework to achieve
the same result by exporting the variable it expects.

Change-Id: I50bbee5a5b23023e40cfc6ef6b372a880ff70001
Signed-off-by: Walinton Cambronero <walinton.cambronero@hpe.com>