Add error detection and better help to query_recipe make target
query-recipe.py only functions properly if devenv is initialized
Split the original dev_header into two parts so that we
can check if devenv is initialized without printing the header
so that the query output is still easily parsable.

Also fixed the error message if the rule is called without any
packages specified.

Change-Id: I627439764d666dcec917face1a7bb8b96f42060f
Signed-off-by: Andrew Crabtree <andrew.crabtree@hpe.com>