patch kernel to support ipv6 ecmp
applying 6 new patches to the linux kernel.
of the list, fix-route-selection-if-kernel-not-complied-with-
CONFIG_IPV6_ROUTER_PREF-modified.patch is the single one that
fixes the round robin behavior of selecting ecmp routes. The
others are either pre-requisite patches in history or important
sounding bug fixes.

This patch set was found by :
1. testing a host runing a 3.19 kernel for ecmp and finding that
 worked.
2. going to linux-cross-reference and comparing differnces between
  3.9 (ours) and 3.18 (close enough to 3.19 and a convient tag)
3. going to git.kernel.org and relating those discoverd differnces
  to exact patches.

In cronological order, the patches are:
https://git.kernel.org/cgit/linux/kernel/git/stable/linux-stable.git/commit/net/ipv6/ip6_fib.c?id=307f2fb95e9b96b3577916e73d92e104f8f26494
https://git.kernel.org/cgit/linux/kernel/git/stable/linux-stable.git/commit/net/ipv6/route.c?id=52bd4c0c1551daa2efa7bb9e01a2f4ea6d1311bb
https://git.kernel.org/cgit/linux/kernel/git/stable/linux-stable.git/commit/net/ipv6/route.c?id=35f1b4e96b9258a3668872b1139c51e5a23eb876
https://git.kernel.org/cgit/linux/kernel/git/stable/linux-stable.git/commit/net/ipv6/route.c?id=35f1b4e96b9258a3668872b1139c51e5a23eb876
https://git.kernel.org/cgit/linux/kernel/git/stable/linux-stable.git/commit/net/ipv6/ip6_fib.c?id=b0e791661b291dfea609f6c647066b005fb546b5
https://git.kernel.org/cgit/linux/kernel/git/stable/linux-stable.git/commit/net/ipv6/route.c?id=27596472473a02cfef2908a6bcda7e55264ba6b7

All but one of those applied cleanly. I had to hand modify afc154e by two
lines before it would apply.

One patch of seeming relevance could not be applied even after
attempting to modifiy the patchfile...
https://git.kernel.org/cgit/linux/kernel/git/stable/linux-stable.git/commit/net/ipv6?id=6b9ea5a64ed5eeb3f68f2e6fcce0ed1179801d1e

Testing done:
  Connect an as5712 to three linux hosts. Confiugre the as5712 to
route to a target-net  via equal cost ipv6 routes through each
of the three nexthops. Ping a range of addresses on the target-net.
telnet to a range of targets on the target-net.

Note: neither the ping, nor the telnets are expected to suceed.
Instead, we watch tcpdump on the nexthops to see where the traffic
is attempting to route through.

Verify that traffic is hashed across the next hops when as5712 has
not previously discoverd the neighbors.

Veirfy that traffic is hashed across the nexthops when as5712 has
previously discovered the neighbors.

Change-Id: I473fd48f5ad10bfc9489164fa1ba251de810206c
Signed-off-by: Ricky Charlet <ricky.charlet@hpe.com>