Parametrize OPS repository information
Instead of having the same information encoded in 53 different places,
add distribution variables that specify the hostname, the path, the
transport protocol and the branch used to fetch OPS code.

In this commit the values are:

	OPS_REPO_HOSTNAME ?= "git.openswitch.net"
	OPS_REPO_PATH ?= "openswitch"
	OPS_REPO_BASE_URL ?= "git://${OPS_REPO_HOSTNAME}/${OPS_REPO_PATH}"
	OPS_REPO_PROTOCOL ?= "https"
	OPS_REPO_BRANCH ?= "master"

In this way switching a branch is a matter of one change in one file.

This cloud be further refined to provide the parameters in the recipe
and generate SRC_URI in openswitch.bbclass (much like python.bbclass
does) but without a clear use case for that it's not worth trying to
implement just yet.

Change-Id: Ife3bf287ce9e33bccd30b3a7762a1841017423f9
Signed-off-by: Marcelo E. Magallon <marcelo.magallon@hpe.com>