ops-ansible.bb: Incorapote test_utils.py component pytest change
This is the SRCREV update to incorporate the test_utils.py
change in ops-ansible.  With this change, We're now using
our own attribute injection file to mount the build ops-ansible
directory onto the ansible control machine docker container
so that we can use the latest playbooks, instead of baking
the playbooks inside the docker image itself.

Tags: dev, chg
Change-Id: I98032ebc4db808c9791a1a74f4a34551f6e3dec1
Signed-off-by: Kei Nohguchi <kei.nohguchi@hpe.com>