fix: dev: Changed the schema-to-idl auto generation script to populate value if it is present.
TG-907
Tags: fix, dev
The auto generation script is not looking for value part of key-value and is
not updating the value in set_subinterface idl api when min=0 and max=1.
Added code to check if value is present fill that also.

Commiting the fix on rel/dill from master, review:
https://review.openswitch.net/#/c/9724/

Resubmitting after merge conflict resolution

Change-Id: I9f91c8f5d562e09247c2b3c24581ead9c70210bd
Signed-off-by: aslammdkhan <aslam.khan@hpe.com>