ops-openvswitch: Add optional C extension wrapper for Python JSON parsing
This requires a refactor the ops-openvswitch recipe.

The python package was not being build/deploy properly,
since we were not calling distutils. This is required in
newer versions of the code that can provide some C versions
of the code for performance reasons.

Adding the change again after was removed due some miss-configuration.

Signed-off-by: Diego Dompe <dompe@hpe.com>

Conflicts:
	yocto/openswitch/meta-distro-openswitch/recipes-ops/openvswitch/ops-openvswitch.bb

Change-Id: I28ecd351620d7ea3a4e4d274970673f9bfbacd00