ops-openvswitch: Add optional C extension wrapper for Python JSON parsing
This requires a refactor the ops-openvswitch recipe.
The python package was not being build/deploy properly,
since we were not calling distutils. This is required in
newer versions of the code that can provide some C versions
of the code for performance reasons.

Submit master commit 1d335cbe1dfa22a0f4649498293b9d2fc23cebf6 to rel/dill

Change-Id: I3cf6d469939aa98f1e16344b758a18bbde6e5ebd
Signed-off-by: Madhulika Madishetty <madhulika.madishetty@hpe.com>