new: dev: Ofproto mirror bundle.
The ofproto_mirror_bundle struct is to enable mirror_configure to pair a
mirror source or destination port with whatever bridge or VRF ofproto it is
currently associated with.
This association of ofprotos with ports allows the PD layer to locate a given
port via it's ofproto number when the mirror is created/modified via mirror_set.

TG-1432

Change-Id: I0867b4c23f7415fdc8b9d97df7abf769f3751308
Signed-off-by: midhan <nitish.midha@hpe.com>
Signed-off-by: Esteban Rodriguez Betancourt <estebarb@hpe.com>