new: dev: Ofproto mirror bundle.
The ofproto_mirror_bundle struct is to enable mirror_configure to pair a
mirror source or destination port with whatever bridge or VRF ofproto it is
currently associated with.
This association of ofprotos with ports allows the PD layer to locate a given
port via it's ofproto number when the mirror is created/modified via mirror_set.

Change-Id: Iec87ac41dd53b008dc31f01c74a0026f0b4e6bba
Signed-off-by: midhan <nitish.midha@hpe.com>
Signed-off-by: Esteban Rodriguez Betancourt <estebarb@hpe.com>