Add support for golang
This adds support for bootstrapping and building Go binaries. The
latest version is 1.6 and thus then one included here.

Since Go is written in Go you also need to bootstrap the 1.6
environment. For this 1.4 is included as this version can be
compiled with gcc.

Thanks to Marcelo Magallon for the golang base.

Change-Id: I1150219b6604efd7947e332c3d947dca4712bf35
Signed-off-by: Christian Svensson <blue@cmd.nu>