Prometheus timeseries monitoring
This adds the prometheus (http://prometheus.io) service to OpenSwitch.
It is available on https://device/prometheus.

Prometheus server is disabled by default but can be enabled by running
'systemctl enable prometheus'.

Prometheus is configured to retain data for 24h or one device reboot,
whichever is sooner. Metrics are kept in tmpfs and are lost on reboot or
power events.

Change-Id: I009a3701dd07aa539b6f1b43342492a9af8e641b
Signed-off-by: Christian Svensson <blue@cmd.nu>