new: dev: Enable OVS tests inside OPS
This patch enables most of the tests from the OVS repo to be run inside
the OPS repo.
Two scripts (run-ovs-tests.sh and ovs-tests.sh) are added that make the
changes needed.
To run the tests just need to execute run-ovs-tests.sh from the
ops-openvswitch folder.
Review the comments inside the scripts to get more information.

Change-Id: I4813aeb3ab15bf34052461b7c91c2639d4ad3ded
Signed-off-by: arnoldo.lutz.guevara@hpe.com <arnoldo.lutz.guevara@hpe.com>
Signed-off-by: Esteban Rodriguez Betancourt <estebarb@hpe.com>