Add more ops dependencies
JSONschema is also needed to build the new modular schema. However, the
original python-jsonschema recipe was not producing
python-jsonschema-native because of the dependencies (all the
dependencies are provided by python-core).

Now, ops depends also on jsonschema and extends the pythonnative recipe,
instead of exporting the PYTHONPATH environment variable.

(.gitignore was updated to exclude the backup files produced by vim)

Change-Id: I4ba4b80203d871b14309de716a913a71617b112a
Signed-off-by: Javier Albornoz <javier.albornoz@hpe.com>