Temporarily remove apidocgen.py from ops-restd gate
Due to the changes needed for ops-restd to support the new schema
modular format, the apidocgen.py execution fails, which prevents the
changes to be merged. This temporarily disables the execution of
apidocgen.py, which will be added back to the recipe once
I232c4e5c9352db24fe349eef3ae1fedd8c8e8f73 is merged.

Tags: chg, dev

Change-Id: Id4847062753ec0eb4545928029aba05a57dabee7
Signed-off-by: Eduardo Ramirez <eramireza@hpe.com>