Support booting the genericx86-64 platform in KVM
This patch adds some kernel tweaks and helper scripts to allow booting
up a local virtual machine running OpenSwitch. This allows to
easier iterate on user flows that are otherwise not testable in Docker.

This also allows testing kernel configurations in an easy way, for
things like AppArmor or capabilities.

To facilitate fast boot times the VirtFS/9P filesystem is used, allowing
the system to boot from an extracted tar ball, just like Docker. This
means that we're reusing all artifacts from the Docker setup.

Change-Id: If7ee3b9894c07b9a8506cbcf4026eb8eeb54145e
Signed-off-by: Christian Svensson <blue@cmd.nu>